# Agent-Runtime-Python/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Runtime API
  version: 1.0.0
  description: Agent Runtime service for LiveKit agents with BitHuman avatar support

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://agent-runtime:8080
    description: Docker service

paths:
  /api/agents/register:
    post:
      summary: Register an agent
      tags: [agents]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAgentRequest'
      responses:
        '200':
          description: Agent registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterAgentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/agents/{agentId}:
    delete:
      summary: Unregister an agent
      tags: [agents]
      security:
        - BearerAuth: []
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Agent unregistered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      summary: Get agent status
      tags: [agents]
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Agent status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStatus'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/agents/:
    get:
      summary: List all agents
      tags: [agents]
      responses:
        '200':
          description: List of agent IDs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/sessions/create:
    post:
      summary: Create a session
      tags: [sessions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/sessions/{sessionId}/end:
    post:
      summary: End a session
      tags: [sessions]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session ended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/sessions/{sessionId}:
    get:
      summary: Get session details
      tags: [sessions]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/metrics/agent/{agentId}:
    get:
      summary: Get agent metrics
      tags: [metrics]
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Agent metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentMetrics'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/metrics/tenant/{tenantId}:
    get:
      summary: Get tenant metrics
      tags: [metrics]
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Tenant metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantMetrics'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/metrics/session/{sessionId}:
    get:
      summary: Get session metrics
      tags: [metrics]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionMetrics'
        '404':
          $ref: '#/components/responses/NotFound'

  /health:
    get:
      summary: Health check
      tags: [health]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /ready:
    get:
      summary: Readiness check
      tags: [health]
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service is not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotReadyResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterAgentRequest:
      type: object
      required: [agentId, tenantId, config]
      properties:
        agentId:
          type: integer
        tenantId:
          type: integer
        config:
          $ref: '#/components/schemas/AgentConfig'

    AgentConfig:
      type: object
      required: [agentId, tenantId, sttProvider, ttsProvider, llmProvider]
      properties:
        agentId:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: string
        sttProvider:
          type: string
        ttsProvider:
          type: string
        llmProvider:
          type: string
        llmModel:
          type: string
        systemPrompt:
          type: string
        visionEnabled:
          type: boolean
        screenShareEnabled:
          type: boolean
        transcribeEnabled:
          type: boolean
        languages:
          type: array
          items:
            type: string
        voiceId:
          type: string
        avatarModel:
          type: string
        maxConcurrentSessions:
          type: integer
        livekitConfig:
          $ref: '#/components/schemas/LiveKitConfig'
        langfuseConfig:
          $ref: '#/components/schemas/LangFuseConfig'

    LiveKitConfig:
      type: object
      required: [url, apiKey, apiSecret]
      properties:
        url:
          type: string
        apiKey:
          type: string
        apiSecret:
          type: string

    LangFuseConfig:
      type: object
      properties:
        enabled:
          type: boolean
        publicKey:
          type: string
        secretKey:
          type: string
        baseUrl:
          type: string

    RegisterAgentResponse:
      type: object
      properties:
        success:
          type: boolean
        agentId:
          type: integer

    AgentStatus:
      type: object
      properties:
        registered:
          type: boolean
        active:
          type: boolean
        activeSessions:
          type: integer
        maxSessions:
          type: integer

    AgentListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            type: integer

    CreateSessionRequest:
      type: object
      required: [agentId, tenantId, roomName]
      properties:
        agentId:
          type: integer
        tenantId:
          type: integer
        roomName:
          type: string
        participantName:
          type: string

    CreateSessionResponse:
      type: object
      properties:
        success:
          type: boolean
        session:
          $ref: '#/components/schemas/SessionInfo'

    SessionInfo:
      type: object
      properties:
        sessionId:
          type: string
        roomName:
          type: string

    Session:
      type: object
      properties:
        sessionId:
          type: string
        agentId:
          type: integer
        tenantId:
          type: integer
        roomName:
          type: string
        status:
          type: string
          enum: [connecting, active, ended, error]
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        participantCount:
          type: integer

    AgentMetrics:
      type: object
      properties:
        agentId:
          type: integer
        totalSessions:
          type: integer
        activeSessions:
          type: integer
        avgLatency:
          type: number
        totalCost:
          type: number

    TenantMetrics:
      type: object
      properties:
        tenantId:
          type: integer
        activeAgents:
          type: integer
        totalSessions:
          type: integer
        totalCost:
          type: number

    SessionMetrics:
      type: object
      properties:
        sessionId:
          type: string
        messageCount:
          type: integer
        avgLatency:
          type: number
        totalCost:
          type: number

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean

    HealthResponse:
      type: object
      properties:
        status:
          type: string
        timestamp:
          type: string
          format: date-time

    NotReadyResponse:
      type: object
      properties:
        status:
          type: string
        reason:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'